# Conflicts:
#	molgenis-one-click-importer/src/main/frontend/src/components/OneClickImporter.vue

+++ molgenis-ontology/src/main/java/org/molgenis/ontology/sorta/job/SortaJobExecution.java
@@ -82,7 +82,7 @@ public class SortaJobExecution extends JobExecution
 
 	public double getThreshold()
 	{
-		return getDouble(THRESHOLD).doubleValue();
+		return getDouble(THRESHOLD);
 	}
 
 	private void setDefaultValues()
